requirements.txt
google-generativeai
google-api-python-client
google-auth-oauthlib
python-dotenv 